Add fixtures

+++ test/test.js
@@ -205,7 +205,7 @@ describe( 'distributions-pareto (type i)-quantile', function tests() {
 			expected,
 			i;
 
-		data = new Float32Array( validationData.data );
+		data = new Float64Array( validationData.data );
 
 		expected = new Float64Array( validationData.expected.map( function( d ) {
 			if (d === 'Inf' ) {
@@ -235,7 +235,7 @@ describe( 'distributions-pareto (type i)-quantile', function tests() {
 			'alpha': validationData.alpha,
 			'beta': validationData.beta
 		});
-		expected = new Float32Array( validationData.expected.map( function( d ) {
+		expected = new Float64Array( validationData.expected.map( function( d ) {
 			if (d === 'Inf' ) {
 				return Number.POSITIVE_INFINITY;
 			}
